roachtest: move monitorImpl to own file
Release note: None
