+++ pkg/cmd/roachtest/BUILD.bazel
@@ -118,6 +118,7 @@ go_library(
         "versionupgrade.go",
         "ycsb.go",
         "z_cluster.go",
+        "z_monitor.go",
         "z_slack.go",
         "z_test_impl.go",
         "z_test_registry.go",
